tests: Add new layer_settings.cpp
