+++ tests/CMakeLists.txt
@@ -133,6 +133,7 @@ target_sources(vk_layer_validation_tests PRIVATE
     unit/instance_positive.cpp
     unit/instanceless.cpp
     unit/layer_utils_positive.cpp
+    unit/layer_settings.cpp
     unit/memory.cpp
     unit/memory_positive.cpp
     unit/mesh.cpp
