+++ tests/unit/others.cpp
@@ -18,22 +18,6 @@
 #include "utils/hash_util.h"
 #include "generated/vk_validation_error_messages.h"
 
-static std::string format(const char *message, ...) {
-    std::size_t const STRING_BUFFER(4096);
-
-    assert(message != nullptr);
-    assert(strlen(message) < STRING_BUFFER);
-
-    char buffer[STRING_BUFFER];
-    va_list list;
-
-    va_start(list, message);
-    vsnprintf(buffer, STRING_BUFFER, message, list);
-    va_end(list);
-
-    return buffer;
-}
-
 TEST_F(VkLayerTest, VersionCheckPromotedAPIs) {
     TEST_DESCRIPTION("Validate that promoted APIs are not valid in old versions.");
     SetTargetApiVersion(VK_API_VERSION_1_0);
@@ -76,163 +60,6 @@ TEST_F(VkLayerTest, UnsupportedPnextApiVersion) {
     }
 }
 
-TEST_F(VkLayerTest, CustomStypeStructString) {
-    TEST_DESCRIPTION("Positive Test for ability to specify custom pNext structs using a list (string)");
-
-    // Create a custom structure
-    typedef struct CustomStruct {
-        VkStructureType sType;
-        const void *pNext;
-        uint32_t custom_data;
-    } CustomStruct;
-
-    uint32_t custom_stype = 3000300000;
-    CustomStruct custom_struct;
-    custom_struct.pNext = nullptr;
-    custom_struct.sType = static_cast<VkStructureType>(custom_stype);
-    custom_struct.custom_data = 44;
-
-    // Communicate list of structinfo pairs to layers
-    const char *id[] = {"3000300000", "24"};
-    const VkLayerSettingEXT setting = {OBJECT_LAYER_NAME, "custom_stype_list", VK_LAYER_SETTING_TYPE_STRING_EXT,
-                                       static_cast<uint32_t>(std::size(id)), &id};
-    VkLayerSettingsCreateInfoEXT layer_setting_create_info = {VK_STRUCTURE_TYPE_LAYER_SETTINGS_CREATE_INFO_EXT, nullptr, 1, &setting};
-
-    RETURN_IF_SKIP(InitFramework(&layer_setting_create_info));
-    RETURN_IF_SKIP(InitState());
-
-    vkt::Buffer buffer(*m_device, 1024, VK_BUFFER_USAGE_STORAGE_TEXEL_BUFFER_BIT);
-    VkBufferViewCreateInfo bvci = vku::InitStructHelper(&custom_struct);  // Add custom struct through pNext
-    bvci.buffer = buffer.handle();
-    bvci.format = VK_FORMAT_R32_SFLOAT;
-    bvci.range = VK_WHOLE_SIZE;
-    vkt::BufferView buffer_view(*m_device, bvci);
-}
-
-TEST_F(VkLayerTest, CustomStypeStructStringArray) {
-    TEST_DESCRIPTION("Positive Test for ability to specify custom pNext structs using a vector of strings");
-
-    // Create a custom structure
-    typedef struct CustomStruct {
-        VkStructureType sType;
-        const void *pNext;
-        uint32_t custom_data;
-    } CustomStruct;
-
-    const uint32_t custom_stype_a = 3000300000;
-    CustomStruct custom_struct_a;
-    custom_struct_a.pNext = nullptr;
-    custom_struct_a.sType = static_cast<VkStructureType>(custom_stype_a);
-    custom_struct_a.custom_data = 44;
-
-    const uint32_t custom_stype_b = 3000300001;
-    CustomStruct custom_struct_b;
-    custom_struct_b.pNext = &custom_struct_a;
-    custom_struct_b.sType = static_cast<VkStructureType>(custom_stype_b);
-    custom_struct_b.custom_data = 88;
-
-    // Communicate list of structinfo pairs to layers, including a duplicate which should get filtered out
-    const std::string string_stype_a = format("%u", custom_stype_a);
-    const std::string string_stype_b = format("%u", custom_stype_b);
-    const std::string sizeof_struct = format("%d", sizeof(CustomStruct));
-
-    const char *ids[] = {
-        string_stype_a.c_str(), sizeof_struct.c_str(),
-        string_stype_b.c_str(), sizeof_struct.c_str(),
-        string_stype_a.c_str(), sizeof_struct.c_str(),
-    };
-    const VkLayerSettingEXT setting = {
-        OBJECT_LAYER_NAME, "custom_stype_list", VK_LAYER_SETTING_TYPE_STRING_EXT, static_cast<uint32_t>(std::size(ids)), &ids};
-    VkLayerSettingsCreateInfoEXT layer_setting_create_info = {
-        VK_STRUCTURE_TYPE_LAYER_SETTINGS_CREATE_INFO_EXT, nullptr, 1, &setting};
-
-    RETURN_IF_SKIP(InitFramework(&layer_setting_create_info));
-    RETURN_IF_SKIP(InitState());
-
-    vkt::Buffer buffer(*m_device, 1024, VK_BUFFER_USAGE_STORAGE_TEXEL_BUFFER_BIT);
-    VkBufferViewCreateInfo bvci = vku::InitStructHelper(&custom_struct_b);  // Add custom struct through pNext
-    bvci.buffer = buffer.handle();
-    bvci.format = VK_FORMAT_R32_SFLOAT;
-    bvci.range = VK_WHOLE_SIZE;
-    vkt::BufferView buffer_view(*m_device, bvci);
-}
-
-TEST_F(VkLayerTest, CustomStypeStructIntegerArray) {
-    TEST_DESCRIPTION("Positive Test for ability to specify custom pNext structs using a vector of integers");
-
-    // Create a custom structure
-    typedef struct CustomStruct {
-        VkStructureType sType;
-        const void *pNext;
-        uint32_t custom_data;
-    } CustomStruct;
-
-    const uint32_t custom_stype_a = 3000300000;
-    CustomStruct custom_struct_a;
-    custom_struct_a.pNext = nullptr;
-    custom_struct_a.sType = static_cast<VkStructureType>(custom_stype_a);
-    custom_struct_a.custom_data = 44;
-
-    const uint32_t custom_stype_b = 3000300001;
-    CustomStruct custom_struct_b;
-    custom_struct_b.pNext = &custom_struct_a;
-    custom_struct_b.sType = static_cast<VkStructureType>(custom_stype_b);
-    custom_struct_b.custom_data = 88;
-
-    // Communicate list of structinfo pairs to layers, including a duplicate which should get filtered out
-    const uint32_t ids[] = {
-        custom_stype_a, sizeof(CustomStruct),
-        custom_stype_b, sizeof(CustomStruct),
-        custom_stype_a, sizeof(CustomStruct)
-    };
-
-    const VkLayerSettingEXT setting[] = {
-        {OBJECT_LAYER_NAME, "custom_stype_list", VK_LAYER_SETTING_TYPE_UINT32_EXT, static_cast<uint32_t>(std::size(ids)), ids}
-    };
-    VkLayerSettingsCreateInfoEXT layer_setting_create_info = {VK_STRUCTURE_TYPE_LAYER_SETTINGS_CREATE_INFO_EXT, nullptr, 1, setting};
-
-    RETURN_IF_SKIP(InitFramework(&layer_setting_create_info));
-    RETURN_IF_SKIP(InitState());
-
-    vkt::Buffer buffer(*m_device, 1024, VK_BUFFER_USAGE_STORAGE_TEXEL_BUFFER_BIT);
-    VkBufferViewCreateInfo bvci = vku::InitStructHelper(&custom_struct_b);  // Add custom struct through pNext
-    bvci.buffer = buffer.handle();
-    bvci.format = VK_FORMAT_R32_SFLOAT;
-    bvci.range = VK_WHOLE_SIZE;
-    vkt::BufferView buffer_view(*m_device, bvci);
-}
-
-TEST_F(VkLayerTest, DuplicateMessageLimit) {
-    TEST_DESCRIPTION("Use the duplicate_message_id setting and verify correct operation");
-    AddRequiredExtensions(VK_KHR_GET_PHYSICAL_DEVICE_PROPERTIES_2_EXTENSION_NAME);
-
-    uint32_t value = 3;
-    const VkLayerSettingEXT setting = {OBJECT_LAYER_NAME, "duplicate_message_limit", VK_LAYER_SETTING_TYPE_UINT32_EXT, 1, &value};
-    VkLayerSettingsCreateInfoEXT create_info = {VK_STRUCTURE_TYPE_LAYER_SETTINGS_CREATE_INFO_EXT, nullptr, 1, &setting};
-
-    RETURN_IF_SKIP(InitFramework(&create_info));
-    RETURN_IF_SKIP(InitState());
-
-    // Create an invalid pNext structure to trigger the stateless validation warning
-    VkBaseOutStructure bogus_struct{};
-    bogus_struct.sType = static_cast<VkStructureType>(0x33333333);
-    VkPhysicalDeviceProperties2KHR properties2 = vku::InitStructHelper(&bogus_struct);
-
-    // Should get the first three errors just fine
-    m_errorMonitor->SetDesiredError("VUID-VkPhysicalDeviceProperties2-pNext-pNext");
-    vk::GetPhysicalDeviceProperties2KHR(Gpu(), &properties2);
-    m_errorMonitor->VerifyFound();
-    m_errorMonitor->SetDesiredError("VUID-VkPhysicalDeviceProperties2-pNext-pNext");
-    vk::GetPhysicalDeviceProperties2KHR(Gpu(), &properties2);
-    m_errorMonitor->VerifyFound();
-    m_errorMonitor->SetDesiredError("VUID-VkPhysicalDeviceProperties2-pNext-pNext");
-    vk::GetPhysicalDeviceProperties2KHR(Gpu(), &properties2);
-    m_errorMonitor->VerifyFound();
-
-    // Limit should prevent the message from coming through a fourth time
-    vk::GetPhysicalDeviceProperties2KHR(Gpu(), &properties2);
-}
-
 TEST_F(VkLayerTest, VuidCheckForHashCollisions) {
     TEST_DESCRIPTION("Ensure there are no VUID hash collisions");
 
@@ -255,104 +82,6 @@ TEST_F(VkLayerTest, VuidHashStability) {
     ASSERT_TRUE(hash_util::VuidHash("VUID-FragDepth-FragDepth-04213") == 0x840af838);
     ASSERT_TRUE(hash_util::VuidHash("VUID-RayTmaxKHR-RayTmaxKHR-04349") == 0x8e67514c);
     ASSERT_TRUE(hash_util::VuidHash("VUID-RuntimeSpirv-SubgroupUniformControlFlowKHR-06379") == 0x2f574188);
-}
-
-TEST_F(VkLayerTest, VuidIdFilterString) {
-    TEST_DESCRIPTION("Validate that message id string filtering is working");
-
-    AddRequiredExtensions(VK_KHR_MAINTENANCE_2_EXTENSION_NAME);
-    // This test would normally produce an unexpected error or two.  Use the message filter instead of
-    // the error_monitor's SetUnexpectedError to test the filtering.
-
-    const char *ids[] = {"VUID-VkRenderPassCreateInfo-pNext-01963"};
-    const VkLayerSettingEXT setting = {OBJECT_LAYER_NAME, "message_id_filter", VK_LAYER_SETTING_TYPE_STRING_EXT, 1, ids};
-    VkLayerSettingsCreateInfoEXT layer_settings_create_info = {VK_STRUCTURE_TYPE_LAYER_SETTINGS_CREATE_INFO_EXT, nullptr, 1, &setting};
-
-    RETURN_IF_SKIP(InitFramework(&layer_settings_create_info));
-
-    RETURN_IF_SKIP(InitState());
-    VkAttachmentDescription attach = {0,
-                                      VK_FORMAT_R8G8B8A8_UNORM,
-                                      VK_SAMPLE_COUNT_1_BIT,
-                                      VK_ATTACHMENT_LOAD_OP_DONT_CARE,
-                                      VK_ATTACHMENT_STORE_OP_DONT_CARE,
-                                      VK_ATTACHMENT_LOAD_OP_DONT_CARE,
-                                      VK_ATTACHMENT_STORE_OP_DONT_CARE,
-                                      VK_IMAGE_LAYOUT_UNDEFINED,
-                                      VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL};
-    VkAttachmentReference ref = {0, VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL};
-    VkSubpassDescription subpass = {0, VK_PIPELINE_BIND_POINT_GRAPHICS, 1, &ref, 0, nullptr, nullptr, nullptr, 0, nullptr};
-    VkInputAttachmentAspectReference iaar = {0, 0, VK_IMAGE_ASPECT_METADATA_BIT};
-    VkRenderPassInputAttachmentAspectCreateInfo rpiaaci = {VK_STRUCTURE_TYPE_RENDER_PASS_INPUT_ATTACHMENT_ASPECT_CREATE_INFO,
-                                                           nullptr, 1, &iaar};
-    VkRenderPassCreateInfo rpci = {VK_STRUCTURE_TYPE_RENDER_PASS_CREATE_INFO, &rpiaaci, 0, 1, &attach, 1, &subpass, 0, nullptr};
-    m_errorMonitor->SetUnexpectedError("VUID-VkRenderPassCreateInfo2-attachment-02525");
-    TestRenderPassCreate(m_errorMonitor, *m_device, rpci, false, "VUID-VkInputAttachmentAspectReference-aspectMask-01964", nullptr);
-}
-
-TEST_F(VkLayerTest, VuidFilterHexInt) {
-    TEST_DESCRIPTION("Validate that message id hex int filtering is working");
-
-    AddRequiredExtensions(VK_KHR_MAINTENANCE_2_EXTENSION_NAME);
-    // This test would normally produce an unexpected error or two.  Use the message filter instead of
-    // the error_monitor's SetUnexpectedError to test the filtering.
-
-    const char *ids[] = {"0xa19880e3"};
-    const VkLayerSettingEXT setting = {OBJECT_LAYER_NAME, "message_id_filter", VK_LAYER_SETTING_TYPE_STRING_EXT, 1, ids};
-    VkLayerSettingsCreateInfoEXT layer_settings_create_info = {VK_STRUCTURE_TYPE_LAYER_SETTINGS_CREATE_INFO_EXT, nullptr, 1, &setting};
-
-    RETURN_IF_SKIP(InitFramework(&layer_settings_create_info));
-
-    RETURN_IF_SKIP(InitState());
-    VkAttachmentDescription attach = {0,
-                                      VK_FORMAT_R8G8B8A8_UNORM,
-                                      VK_SAMPLE_COUNT_1_BIT,
-                                      VK_ATTACHMENT_LOAD_OP_DONT_CARE,
-                                      VK_ATTACHMENT_STORE_OP_DONT_CARE,
-                                      VK_ATTACHMENT_LOAD_OP_DONT_CARE,
-                                      VK_ATTACHMENT_STORE_OP_DONT_CARE,
-                                      VK_IMAGE_LAYOUT_UNDEFINED,
-                                      VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL};
-    VkAttachmentReference ref = {0, VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL};
-    VkSubpassDescription subpass = {0, VK_PIPELINE_BIND_POINT_GRAPHICS, 1, &ref, 0, nullptr, nullptr, nullptr, 0, nullptr};
-    VkInputAttachmentAspectReference iaar = {0, 0, VK_IMAGE_ASPECT_METADATA_BIT};
-    VkRenderPassInputAttachmentAspectCreateInfo rpiaaci = {VK_STRUCTURE_TYPE_RENDER_PASS_INPUT_ATTACHMENT_ASPECT_CREATE_INFO,
-                                                           nullptr, 1, &iaar};
-    VkRenderPassCreateInfo rpci = {VK_STRUCTURE_TYPE_RENDER_PASS_CREATE_INFO, &rpiaaci, 0, 1, &attach, 1, &subpass, 0, nullptr};
-    m_errorMonitor->SetUnexpectedError("VUID-VkRenderPassCreateInfo2-attachment-02525");
-    TestRenderPassCreate(m_errorMonitor, *m_device, rpci, false, "VUID-VkInputAttachmentAspectReference-aspectMask-01964", nullptr);
-}
-
-TEST_F(VkLayerTest, VuidFilterInt) {
-    TEST_DESCRIPTION("Validate that message id decimal int filtering is working");
-
-    AddRequiredExtensions(VK_KHR_MAINTENANCE_2_EXTENSION_NAME);
-    // This test would normally produce an unexpected error or two.  Use the message filter instead of
-    // the error_monitor's SetUnexpectedError to test the filtering.
-
-    const char *ids[] = {"2711126243"};
-    const VkLayerSettingEXT setting = {OBJECT_LAYER_NAME, "message_id_filter", VK_LAYER_SETTING_TYPE_STRING_EXT, 1, ids};
-    VkLayerSettingsCreateInfoEXT layer_settings_create_info = {VK_STRUCTURE_TYPE_LAYER_SETTINGS_CREATE_INFO_EXT, nullptr, 1, &setting};
-
-    RETURN_IF_SKIP(InitFramework(&layer_settings_create_info));
-    RETURN_IF_SKIP(InitState());
-    VkAttachmentDescription attach = {0,
-                                      VK_FORMAT_R8G8B8A8_UNORM,
-                                      VK_SAMPLE_COUNT_1_BIT,
-                                      VK_ATTACHMENT_LOAD_OP_DONT_CARE,
-                                      VK_ATTACHMENT_STORE_OP_DONT_CARE,
-                                      VK_ATTACHMENT_LOAD_OP_DONT_CARE,
-                                      VK_ATTACHMENT_STORE_OP_DONT_CARE,
-                                      VK_IMAGE_LAYOUT_UNDEFINED,
-                                      VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL};
-    VkAttachmentReference ref = {0, VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL};
-    VkSubpassDescription subpass = {0, VK_PIPELINE_BIND_POINT_GRAPHICS, 1, &ref, 0, nullptr, nullptr, nullptr, 0, nullptr};
-    VkInputAttachmentAspectReference iaar = {0, 0, VK_IMAGE_ASPECT_METADATA_BIT};
-    VkRenderPassInputAttachmentAspectCreateInfo rpiaaci = {VK_STRUCTURE_TYPE_RENDER_PASS_INPUT_ATTACHMENT_ASPECT_CREATE_INFO,
-                                                           nullptr, 1, &iaar};
-    VkRenderPassCreateInfo rpci = {VK_STRUCTURE_TYPE_RENDER_PASS_CREATE_INFO, &rpiaaci, 0, 1, &attach, 1, &subpass, 0, nullptr};
-    m_errorMonitor->SetUnexpectedError("VUID-VkRenderPassCreateInfo2-attachment-02525");
-    TestRenderPassCreate(m_errorMonitor, *m_device, rpci, false, "VUID-VkInputAttachmentAspectReference-aspectMask-01964", nullptr);
 }
 
 TEST_F(VkLayerTest, RequiredParameter) {
